Support for trait: throw_missing
This avoids the exists/get pair of roundtrips and replaces them with
a single get | get/fail approach.  This is supported for RedisApi based
drivers because the serialized data and the key will never be NULL, and
that is what will be returned when missing.  It is not implemented for
rds (even though it could be) because file.exists() is much faster than
catching a failure in readLines().