Overhaul large file support
For #25

The hack for getting long vectors serialised was slow and consumed too much
memory.  This new hack is also slow (and still consumes a lot of memory) but
is _less_ slow than the other approaches that I could come up with.  Still,
this seems like a far from ideal approach for storing very large data.

Once R supports long vectors with writeBin, things will be a bit faster and
consume less memory.