Allow import/export of multiple namespaces at once
With the mget/mset support this could involve relatively few
roundtrips as the whole interchange is done at once.  This could
cause things to run out of memory though, and I'll need to see
what can be done to batch this up (in general it would require
that storr drivers could report back the size of objects and that's
nasty, so we could just offer a one-by-one option).